commit fixed 6-size.c error

+++ 0x00-hello_world/6-size.c
@@ -10,10 +10,10 @@ long int g;
 long long int t;
 char r;
 float p;
-printf("Size of a char: %d byte(s)\n", sizeof(r));
-printf("Size of an int: %d byte(s)\n", sizeof(d));
-printf("Size of long int: %d byte(s)\n", sizeof(g));
-printf("Size of long long int: %d byte(s)\n", sizeof(t));
-printf("Size of float: %d byte(s)\n", sizeof(p));
+printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(r));
+printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(d));
+printf("Size of long int: %lu byte(s)\n", (unsigned long)sizeof(g));
+printf("Size of long long int: %lu byte(s)\n", (unsigned long)sizeof(t));
+printf("Size of float: %lu byte(s)\n", (unsigned long)sizeof(p));
 return (0);
 }
